add role and separate seed_prompt_group and seed_prompt_dataset
