+++ pyrit/datasets/seed_prompts/illegal-multimodal-dataset.prompt
@@ -1,4 +1,4 @@
-dataset_name: 2025_06_pyrit_illegal_multimodal_example
+dataset_name: 2025_06_pyrit_illegal_multimodal_multiturn_example
 harm_categories: illegal
 source: https://azure.github.io/PyRIT/
 groups: AI Red Team
