+++ pyrit/models/seed_prompt.py
@@ -6,17 +6,14 @@ from __future__ import annotations
 import logging
 import os
 import uuid
-from collections import defaultdict
 from dataclasses import dataclass, field
 from datetime import datetime
 from pathlib import Path
-from typing import Any, Dict, Optional, Sequence, Union
+from typing import Dict, Optional, Sequence, Union
 
 from jinja2 import BaseLoader, Environment, StrictUndefined, Template, Undefined
-from pydantic.types import PositiveInt
 from tinytag import TinyTag
 
-from pyrit.common import utils
 from pyrit.common.path import PATHS_DICT
 from pyrit.common.yaml_loadable import YamlLoadable
 from pyrit.models import DataTypeSerializer
@@ -45,23 +42,61 @@ class PartialUndefined(Undefined):
 class SeedPrompt(YamlLoadable):
     """Represents a seed prompt with various attributes and metadata."""
 
+    # The actual prompt value, which can be a string or a file path
     value: str
+
+    # SHA256 hash of the value, used for deduplication
     value_sha256: Optional[str] = None
+
+    # The type of data this prompt represents (e.g., text, image, audio, video)
     data_type: Optional[PromptDataType] = None
+
+    # Unique identifier for the prompt
     id: Optional[uuid.UUID] = field(default_factory=lambda: uuid.uuid4())
+
+    # Name of the prompt
     name: Optional[str] = None
+
+    # Name of the dataset this prompt belongs to
     dataset_name: Optional[str] = None
+
+    # Categories of harm associated with this prompt
     harm_categories: Optional[Sequence[str]] = field(default_factory=lambda: [])
+
+    # Description of the prompt
     description: Optional[str] = None
+
+    # Authors of the prompt
     authors: Optional[Sequence[str]] = field(default_factory=lambda: [])
+
+    # Groups affiliated with the prompt
     groups: Optional[Sequence[str]] = field(default_factory=lambda: [])
+
+    # Source of the prompt
     source: Optional[str] = None
+
+    # Date when the prompt was added to the dataset
     date_added: Optional[datetime] = field(default_factory=lambda: datetime.now())
+
+    # User who added the prompt to the dataset
     added_by: Optional[str] = None
+
+    # Arbitrary metadata that can be attached to the prompt
     metadata: Optional[Dict[str, Union[str, int]]] = field(default_factory=lambda: {})
+
+    # Parameters that can be used in the prompt template
     parameters: Optional[Sequence[str]] = field(default_factory=lambda: [])
+
+    # Unique identifier for the prompt group
     prompt_group_id: Optional[uuid.UUID] = None
+    
+    # Alias for the prompt group
     prompt_group_alias: Optional[str] = None
+
+    # Role of the prompt in a conversation (e.g., "user", "assistant")
+    role: Optional[str] = "user"
+
+    # Sequence number for ordering prompts in a conversation
     sequence: Optional[int] = 0
 
     def __post_init__(self) -> None:
@@ -209,292 +244,3 @@ class SeedPrompt(YamlLoadable):
 
         return sp
 
-
-class SeedPromptGroup(YamlLoadable):
-    """
-    A group of prompts that need to be sent together, along with an objective.
-
-    This class is useful when a target requires multiple (multimodal) prompt pieces to be grouped
-    and sent together. All prompts in the group should share the same `prompt_group_id`.
-    """
-
-    prompts: Sequence[SeedPrompt]
-
-    def __init__(
-        self,
-        *,
-        prompts: Union[Sequence[SeedPrompt], Sequence[Dict[str, Any]]],
-    ):
-        if not prompts:
-            raise ValueError("SeedPromptGroup cannot be empty.")
-        self.prompts = []
-        for prompt in prompts:
-            if isinstance(prompt, SeedPrompt):
-                self.prompts.append(prompt)
-            elif isinstance(prompt, dict):
-                self.prompts.append(SeedPrompt(**prompt))
-
-        self._enforce_consistent_group_id()
-
-        # Check sequence and sort the prompts in the same loop
-        if len(self.prompts) >= 1:
-            self.prompts = sorted(
-                self.prompts, key=lambda prompt: prompt.sequence if prompt.sequence is not None else 0
-            )
-
-    def render_template_value(self, **kwargs):
-        """Renders self.value as a template, applying provided parameters in kwargs
-
-        Args:
-            kwargs:Key-value pairs to replace in the SeedPromptGroup value.
-
-        Returns:
-            None
-
-        Raises:
-            ValueError: If parameters are missing or invalid in the template.
-        """
-
-        for prompt in self.prompts:
-            prompt.value = prompt.render_template_value(**kwargs)
-
-    def _enforce_consistent_group_id(self):
-        """
-        Ensures that if any of the prompts already have a group ID set,
-        they share the same ID. If none have a group ID set, assign a
-        new UUID to all prompts.
-
-        Raises:
-            ValueError: If multiple different group IDs exist among the prompts.
-        """
-        existing_group_ids = {prompt.prompt_group_id for prompt in self.prompts if prompt.prompt_group_id is not None}
-
-        if len(existing_group_ids) > 1:
-            # More than one distinct group ID found among prompts.
-            raise ValueError("Inconsistent group IDs found across prompts.")
-        elif len(existing_group_ids) == 1:
-            # Exactly one group ID is set; apply it to all.
-            group_id = existing_group_ids.pop()
-            for prompt in self.prompts:
-                prompt.prompt_group_id = group_id
-        else:
-            # No group IDs set; generate a fresh one and assign it to all.
-            new_group_id = uuid.uuid4()
-            for prompt in self.prompts:
-                prompt.prompt_group_id = new_group_id
-
-    def is_single_request(self) -> bool:
-        unique_sequences = {prompt.sequence for prompt in self.prompts}
-        return len(unique_sequences) <= 1
-
-    def is_single_part_single_text_request(self) -> bool:
-        return len(self.prompts) == 1 and self.prompts[0].data_type == "text"
-
-    def __repr__(self):
-        return f"<SeedPromptGroup(prompts={len(self.prompts)} prompts)>"
-
-
-class SeedPromptDataset(YamlLoadable):
-    """
-    SeedPromptDataset manages seed prompts plus optional top-level defaults.
-    Prompts are stored as a Sequence[SeedPrompt], so references to prompt properties
-    are straightforward (e.g. ds.prompts[0].value).
-    """
-
-    data_type: Optional[str]
-    name: Optional[str]
-    dataset_name: Optional[str]
-    harm_categories: Optional[Sequence[str]]
-    description: Optional[str]
-    authors: Optional[Sequence[str]]
-    groups: Optional[Sequence[str]]
-    source: Optional[str]
-    date_added: Optional[datetime]
-    added_by: Optional[str]
-
-    # Now the actual prompts
-    prompts: Sequence["SeedPrompt"]
-
-    def __init__(
-        self,
-        *,
-        prompts: Optional[Union[Sequence[Dict[str, Any]], Sequence[SeedPrompt]]] = None,
-        data_type: Optional[PromptDataType] = "text",
-        name: Optional[str] = None,
-        dataset_name: Optional[str] = None,
-        harm_categories: Optional[Sequence[str]] = None,
-        description: Optional[str] = None,
-        authors: Optional[Sequence[str]] = None,
-        groups: Optional[Sequence[str]] = None,
-        source: Optional[str] = None,
-        date_added: Optional[datetime] = None,
-        added_by: Optional[str] = None,
-    ):
-        """
-        Initialize the dataset.
-        Typically, you'll call from_dict or from_yaml_file so that top-level defaults
-        are merged into each prompt. If you're passing prompts directly, they can be
-        either a list of SeedPrompt objects or prompt dictionaries (which then get
-        converted to SeedPrompt objects).
-        """
-        if prompts is None:
-            prompts = []
-        if not prompts:
-            raise ValueError("SeedPromptDataset cannot be empty.")
-
-        # Store top-level fields
-        self.data_type = data_type
-        self.name = name
-        self.dataset_name = dataset_name
-
-        self.harm_categories = harm_categories
-        self.description = description
-        self.authors = authors or []
-        self.groups = groups or []
-        self.source = source
-        self.date_added = date_added or datetime.now()
-        self.added_by = added_by
-
-        # Convert any dictionaries in `prompts` to SeedPrompt objects
-        self.prompts = []
-        for p in prompts:
-            if isinstance(p, dict):
-                self.prompts.append(SeedPrompt(**p))
-            elif isinstance(p, SeedPrompt):
-                self.prompts.append(p)
-            else:
-                raise ValueError("Prompts should be either dicts or SeedPrompt objects. Got something else.")
-
-    def get_values(self, first: Optional[PositiveInt] = None, last: Optional[PositiveInt] = None) -> Sequence[str]:
-        """
-        Extracts and returns a list of prompt values from the dataset. By default, returns all of them.
-
-        Args:
-            first (Optional[int]): If provided, values from the first N prompts are included.
-            last (Optional[int]): If provided, values from the last N prompts are included.
-
-        Returns:
-            Sequence[str]: A list of prompt values.
-        """
-        values = [prompt.value for prompt in self.prompts]
-
-        if first is None and last is None:
-            return values
-        if first and last and first + last >= len(values):
-            return values  # simply return all values in case of an overlap
-
-        first_part = values[:first] if first is not None else []
-        last_part = values[-last:] if last is not None else []
-
-        return first_part + last_part
-
-    @classmethod
-    def from_dict(cls, data: Dict[str, Any]) -> "SeedPromptDataset":
-        """
-        Builds a SeedPromptDataset by merging top-level defaults into each item in 'prompts'.
-        """
-        # Pop out the prompts section
-        prompts_data = data.pop("prompts", [])
-        dataset_defaults = data  # everything else is top-level
-
-        merged_prompts = []
-        for p in prompts_data:
-            # Merge dataset-level fields with the prompt-level fields
-            merged = utils.combine_dict(dataset_defaults, p)
-
-            merged["harm_categories"] = utils.combine_list(
-                dataset_defaults.get("harm_categories", []),
-                p.get("harm_categories", []),
-            )
-
-            merged["authors"] = utils.combine_list(
-                dataset_defaults.get("authors", []),
-                p.get("authors", []),
-            )
-
-            merged["groups"] = utils.combine_list(
-                dataset_defaults.get("groups", []),
-                p.get("groups", []),
-            )
-
-            if "data_type" not in merged:
-                merged["data_type"] = dataset_defaults.get("data_type", "text")
-
-            merged_prompts.append(merged)
-
-        for prompt in merged_prompts:
-            if "prompt_group_id" in prompt:
-                raise ValueError("prompt_group_id should not be set in prompt data")
-
-        SeedPromptDataset._set_seed_prompt_group_id_by_alias(seed_prompts=merged_prompts)
-
-        # Now create the dataset with the newly merged prompt dicts
-        return cls(prompts=merged_prompts, **dataset_defaults)
-
-    def render_template_value(self, **kwargs):
-        """Renders self.value as a template, applying provided parameters in kwargs
-
-        Args:
-            kwargs:Key-value pairs to replace in the SeedPromptDataset value.
-
-        Returns:
-            None
-
-        Raises:
-            ValueError: If parameters are missing or invalid in the template.
-        """
-
-        for prompt in self.prompts:
-            prompt.value = prompt.render_template_value(**kwargs)
-
-    @staticmethod
-    def _set_seed_prompt_group_id_by_alias(seed_prompts: Sequence[dict]):
-        """
-        Sets all seed_prompt_group_ids based on prompt_group_id_alias matches
-
-        This is important so the prompt_group_id_alias can be set in yaml to group prompts
-        """
-        alias_to_group_id = {}
-
-        for prompt in seed_prompts:
-            alias = prompt.get("prompt_group_alias")
-            if alias:
-                if alias not in alias_to_group_id:
-                    alias_to_group_id[alias] = uuid.uuid4()
-                prompt["prompt_group_id"] = alias_to_group_id[alias]
-            else:
-                prompt["prompt_group_id"] = uuid.uuid4()
-
-    @staticmethod
-    def group_seed_prompts_by_prompt_group_id(seed_prompts: Sequence[SeedPrompt]) -> Sequence[SeedPromptGroup]:
-        """
-        Groups the given list of SeedPrompts by their prompt_group_id and creates
-        SeedPromptGroup instances.
-
-        Args:
-            seed_prompts: A list of SeedPrompt objects.
-
-        Returns:
-            A list of SeedPromptGroup objects, with prompts grouped by prompt_group_id.
-        """
-        # Group seed prompts by `prompt_group_id`
-        grouped_prompts = defaultdict(list)
-        for prompt in seed_prompts:
-            if prompt.prompt_group_id:
-                grouped_prompts[prompt.prompt_group_id].append(prompt)
-            else:
-                grouped_prompts[uuid.uuid4()].append(prompt)
-
-        # Create SeedPromptGroup instances from grouped prompts
-        seed_prompt_groups = []
-        for group_prompts in grouped_prompts.values():
-            if len(group_prompts) > 1:
-                group_prompts.sort(key=lambda prompt: prompt.sequence)
-
-            seed_prompt_group = SeedPromptGroup(prompts=group_prompts)
-            seed_prompt_groups.append(seed_prompt_group)
-
-        return seed_prompt_groups
-
-    def __repr__(self):
-        return f"<SeedPromptDataset(prompts={len(self.prompts)} prompts)>"
