+++ pyrit/prompt_normalizer/prompt_normalizer.py
@@ -16,7 +16,7 @@ from pyrit.models import (
     construct_response_from_request,
 )
 from pyrit.models.filter_criteria import PromptConverterState, PromptFilterCriteria
-from pyrit.models.seed_prompt import SeedPromptGroup
+from pyrit.models.seed_prompt_group import SeedPromptGroup
 from pyrit.prompt_normalizer import NormalizerRequest, PromptConverterConfiguration
 from pyrit.prompt_target import PromptTarget
 from pyrit.prompt_target.batch_helper import batch_task_async
@@ -47,7 +47,6 @@ class PromptNormalizer:
         conversation_id: Optional[str] = None,
         request_converter_configurations: list[PromptConverterConfiguration] = [],
         response_converter_configurations: list[PromptConverterConfiguration] = [],
-        sequence: int = -1,
         labels: Optional[dict[str, str]] = None,
         orchestrator_identifier: Optional[dict[str, str]] = None,
     ) -> PromptRequestResponse:
@@ -62,7 +61,6 @@ class PromptNormalizer:
                 converting the request. Defaults to an empty list.
             response_converter_configurations (list[PromptConverterConfiguration], optional): Configurations for
                 converting the response. Defaults to an empty list.
-            sequence (int, optional): The sequence number of the request. Defaults to -1.
             labels (Optional[dict[str, str]], optional): Labels associated with the request. Defaults to None.
             orchestrator_identifier (Optional[dict[str, str]], optional): Identifier for the orchestrator. Defaults to
                 None.
@@ -79,7 +77,6 @@ class PromptNormalizer:
             conversation_id=conversation_id,
             request_converter_configurations=request_converter_configurations,
             target=target,
-            sequence=sequence,
             labels=labels,
             orchestrator_identifier=orchestrator_identifier,
         )
@@ -301,7 +298,6 @@ class PromptNormalizer:
         conversation_id: str,
         request_converter_configurations: list[PromptConverterConfiguration],
         target: PromptTarget,
-        sequence: int,
         labels: dict[str, str],
         orchestrator_identifier: Optional[dict[str, str]] = None,
     ) -> PromptRequestResponse:
@@ -316,7 +312,6 @@ class PromptNormalizer:
             request_converter_configurations (list[PromptConverterConfiguration]): List of configurations for
                 request converters.
             target (PromptTarget): The target for the prompt.
-            sequence (int): The sequence number of the prompt.
             labels (dict[str, str]): A dictionary of labels associated with the prompt.
             orchestrator_identifier (Optional[dict[str, str]]): An optional dictionary for orchestrator identifiers.
 
@@ -329,12 +324,11 @@ class PromptNormalizer:
         # All prompt request pieces within PromptRequestResponse needs to have same conversation ID.
         conversation_id = conversation_id if conversation_id else str(uuid4())
         for seed_prompt in seed_prompt_group.prompts:
-
             prompt_request_piece = PromptRequestPiece(
-                role="user",
+                role=seed_prompt.role,
                 original_value=seed_prompt.value,
                 conversation_id=conversation_id,
-                sequence=sequence,
+                sequence=seed_prompt.sequence,
                 labels=labels,
                 prompt_metadata=seed_prompt.metadata,
                 prompt_target_identifier=target.get_identifier(),
@@ -347,6 +341,8 @@ class PromptNormalizer:
         response = PromptRequestResponse(request_pieces=entries)
 
         await self.convert_values(converter_configurations=request_converter_configurations, request_response=response)
+        for r in response.request_pieces:
+            print(f"Request piece: {r.original_value} (sequence: {r.sequence}) (role: {r.role})")
         return response
 
     async def add_prepended_conversation_to_memory(
